fix: Fixes writing to dir when no build has ran

+++ vite.config.ts
@@ -9,21 +9,25 @@ export default defineConfig({
   plugins: [
     dts(),
     () => {
-      fs.copyFileSync('README.md', './dist/README.md')
+      if (fs.existsSync('dist')) {
+        fs.copyFileSync('README.md', 'dist/README.md')
+      }
     },
     () => {
-      const packageDeployable = {}
+      if (fs.existsSync('dist')) {
+        const packageDeployable = {}
 
-      for (const [key, value] of Object.entries(pkg)) {
-        if (!['dependencies', 'devDependencies', 'scripts'].includes(key)) {
-          packageDeployable[key] = value
+        for (const [key, value] of Object.entries(pkg)) {
+          if (!['dependencies', 'devDependencies', 'scripts'].includes(key)) {
+            packageDeployable[key] = value
+          }
         }
-      }
 
-      fs.writeFileSync(
-        './dist/package.json',
-        JSON.stringify(packageDeployable, undefined, 2)
-      )
+        fs.writeFileSync(
+          './dist/package.json',
+          JSON.stringify(packageDeployable, undefined, 2)
+        )
+      }
     },
   ],
   build: {
